limit to Int64

+++ test/fwds_rvs_data.jl
@@ -25,10 +25,10 @@ end
     end
     @test_throws ErrorException(
         "Int64 is a primitive type. Implement a method of `rdata_type` for it."
-    ) rdata_type(Int)
+    ) rdata_type(Int64)
     @test_throws ErrorException(
         "Int64 is a primitive type. Implement a method of `fdata_type` for it."
-    ) fdata_type(Int)
+    ) fdata_type(Int64)
 
     @testset "$(typeof(p))" for (_, p, _...) in Mooncake.tangent_test_cases()
         TestUtils.test_tangent_splitting(Xoshiro(123456), p)
